feat(map): add taint analysis to XHR resolver; extract shared taint utilities
Moves taint analysis infrastructure (inferEnclosingFn, resolveParamProperty,
substituteCallerPlaceholders, substituteCallerHeaders, makeGetCallers) out of
vue_resolveFetch into a new taint_utils.ts module so both the fetch and XHR
resolvers share the same caller-tracing logic.

Extends substituteCallerPlaceholders with a [param:P] case: when the URL is
a direct function parameter (not a property of one), callers are searched for
what value was passed as the first argument and substituted in-place.

vue_resolveXhr now captures the enclosing function for each .open() call,
builds a caller lookup over all JS files, and runs a second pass substituting
[param:X], [member:P.X], [urlsearchparams:P.X], and [call:X.toUpperCase()]
placeholders after the initial AST scan.
