+++ src/map/vue_js/vue_resolveFetch.ts
@@ -3,8 +3,15 @@ import parser from "@babel/parser";
 import _traverse from "@babel/traverse";
 import fs from "fs";
 import path from "path";
-import { resolveNodeValue, substituteVariablesInString, memberChainToString } from "../next_js/utils.js";
+import { resolveNodeValue, substituteVariablesInString } from "../next_js/utils.js";
 import * as globals from "../../utility/globals.js";
+import {
+    EnclosingFn,
+    inferEnclosingFn,
+    substituteCallerPlaceholders,
+    substituteCallerHeaders,
+    makeGetCallers,
+} from "./taint_utils.js";
 
 const traverse = _traverse.default;
 
@@ -85,24 +92,6 @@ const isFunctionLike = (node: any): boolean =>
         node.type === "FunctionExpression" ||
         node.type === "FunctionDeclaration");
 
-interface EnclosingFn {
-    bindingName: string | null;
-    firstParamName: string | null;
-    node: any;
-    file: string;
-}
-
-interface CallerInfo {
-    file: string;
-    fileContent: string;
-    callNode: any;
-    scope: any;
-    args: any[];
-    // Function that contains this call — used for transitive resolution when
-    // the call's argument is itself the enclosing function's parameter.
-    enclosingFn: EnclosingFn | null;
-}
-
 interface FetchEntry {
     file: string;
     filePath: string;
@@ -115,306 +104,6 @@ interface FetchEntry {
     enclosingFn: EnclosingFn | null;
 }
 
-/**
- * Pulls out the function binding name from the surrounding declarator /
- * assignment so we can later match callsites by identifier. Handles four
- * common shapes seen in Vite-bundled code:
- *   const fn = (arg) => { ... }
- *   fn = (arg) => { ... }
- *   function fn(arg) { ... }
- *   { fn: (arg) => { ... } }  // object-literal property value
- */
-const inferEnclosingFn = (callPath: any, file: string): EnclosingFn | null => {
-    const fnPath = callPath.getFunctionParent();
-    if (!fnPath) return null;
-    const fnNode = fnPath.node;
-    const firstParamName =
-        fnNode.params && fnNode.params[0]?.type === "Identifier"
-            ? fnNode.params[0].name
-            : fnNode.params &&
-                fnNode.params[0]?.type === "AssignmentPattern" &&
-                fnNode.params[0].left?.type === "Identifier"
-              ? fnNode.params[0].left.name
-              : null;
-
-    let bindingName: string | null = null;
-    if (fnNode.type === "FunctionDeclaration" && fnNode.id?.type === "Identifier") {
-        bindingName = fnNode.id.name;
-    } else {
-        const parentNode = fnPath.parentPath?.node;
-        if (parentNode) {
-            if (parentNode.type === "VariableDeclarator" && parentNode.id?.type === "Identifier") {
-                bindingName = parentNode.id.name;
-            } else if (parentNode.type === "AssignmentExpression" && parentNode.left?.type === "Identifier") {
-                bindingName = parentNode.left.name;
-            } else if (
-                parentNode.type === "ObjectProperty" &&
-                !parentNode.computed &&
-                parentNode.key?.type === "Identifier"
-            ) {
-                bindingName = parentNode.key.name;
-            }
-        }
-    }
-
-    return { bindingName, firstParamName, node: fnNode, file };
-};
-
-/**
- * Walks an ObjectExpression and returns the property value for the given
- * dotted property path (e.g. ["data"] -> the value node of `data: ...`).
- * Returns null if any segment of the path isn't a literal property.
- */
-const lookupObjectExpressionProp = (objExpr: any, propPath: string[]): any => {
-    if (!objExpr || objExpr.type !== "ObjectExpression") return null;
-    let current: any = objExpr;
-    for (const segment of propPath) {
-        if (!current || current.type !== "ObjectExpression") return null;
-        let next: any = null;
-        for (const prop of current.properties) {
-            if (prop.type !== "ObjectProperty") continue;
-            const key =
-                prop.key.type === "Identifier"
-                    ? prop.key.name
-                    : prop.key.type === "StringLiteral"
-                      ? prop.key.value
-                      : null;
-            if (key === segment) {
-                next = prop.value;
-                break;
-            }
-        }
-        if (!next) return null;
-        current = next;
-    }
-    return current;
-};
-
-/**
- * Resolves the value of `paramName.propPath` by walking back through the
- * enclosing function's callers. If a caller's argument is itself an Identifier
- * pointing to a constant initializer, we follow that binding; if it's the
- * caller's own first parameter, we recurse to that function's callers.
- *
- * Returns the resolved AST node + the scope it lives in (so further sub-property
- * resolution can use the correct binding lookup), or null when the chain breaks.
- */
-const resolveParamProperty = (
-    paramName: string,
-    propPath: string[],
-    enclosingFn: EnclosingFn | null,
-    getCallers: (bindingName: string) => CallerInfo[],
-    depth: number = 0
-): { node: any; scope: any; fileContent: string } | null => {
-    if (!enclosingFn || !enclosingFn.bindingName || depth > 6) return null;
-    // Only the function's first parameter is supported for now — every observed
-    // case has fetch wrappers shaped as `(e) => fetch(... e.X ...)`.
-    if (enclosingFn.firstParamName !== paramName) return null;
-
-    const callers = getCallers(enclosingFn.bindingName);
-    if (!callers || callers.length === 0) return null;
-
-    for (const caller of callers) {
-        const arg = caller.args[0];
-        if (!arg) continue;
-
-        // Case 1: caller passes an object literal directly — fn({ a: ..., b: ... })
-        if (arg.type === "ObjectExpression") {
-            const node = lookupObjectExpressionProp(arg, propPath);
-            if (node) return { node, scope: caller.scope, fileContent: caller.fileContent };
-        }
-
-        // Case 2: caller passes an Identifier pointing to a const-initialized
-        // object literal — `const x = { a: ... }; fn(x)`.
-        if (arg.type === "Identifier") {
-            const binding = caller.scope.getBinding(arg.name);
-            const initNode = binding?.path?.node?.init;
-            if (initNode && initNode.type === "ObjectExpression") {
-                const node = lookupObjectExpressionProp(initNode, propPath);
-                if (node) return { node, scope: caller.scope, fileContent: caller.fileContent };
-            }
-            // Case 3: caller passes its own first param straight through —
-            // `outer = (p) => fn(p)`. Recurse up to that function's callers.
-            if (caller.enclosingFn?.firstParamName === arg.name) {
-                const result = resolveParamProperty(arg.name, propPath, caller.enclosingFn, getCallers, depth + 1);
-                if (result) return result;
-            }
-        }
-    }
-    return null;
-};
-
-/**
- * Renders an ObjectExpression as a `k1={k1}&k2={k2}` query-string fragment.
- * Literal property values are URL-encoded directly; non-literal values fall
- * back to the placeholder form so the schema reader can see what's missing.
- */
-const renderObjectAsQuery = (objExpr: any, scope: any, fileContent: string): string | null => {
-    if (!objExpr || objExpr.type !== "ObjectExpression") return null;
-    const parts: string[] = [];
-    for (const prop of objExpr.properties) {
-        if (prop.type !== "ObjectProperty") continue;
-        const key =
-            prop.key.type === "Identifier" ? prop.key.name : prop.key.type === "StringLiteral" ? prop.key.value : null;
-        if (!key) continue;
-        let valStr: string;
-        try {
-            const resolved = resolveNodeValue(prop.value, scope, "", "fetch", fileContent);
-            if (resolved !== null && resolved !== undefined && !String(resolved).startsWith("[")) {
-                valStr = encodeURIComponent(String(resolved));
-            } else {
-                valStr = `{${key}}`;
-            }
-        } catch {
-            valStr = `{${key}}`;
-        }
-        parts.push(`${key}=${valStr}`);
-    }
-    return parts.length > 0 ? parts.join("&") : null;
-};
-
-/**
- * Returns a plain object derived from an ObjectExpression's properties.
- * Spread elements are merged in when they themselves resolve to a literal
- * object; otherwise the spread is preserved as a sentinel key so the schema
- * reader knows extra fields exist at runtime.
- */
-const renderObjectExpression = (objExpr: any, scope: any, fileContent: string): Record<string, string> | null => {
-    if (!objExpr || objExpr.type !== "ObjectExpression") return null;
-    const out: Record<string, string> = {};
-    for (const prop of objExpr.properties) {
-        if (prop.type === "ObjectProperty") {
-            const key =
-                prop.key.type === "Identifier"
-                    ? prop.key.name
-                    : prop.key.type === "StringLiteral"
-                      ? prop.key.value
-                      : null;
-            if (!key) continue;
-            try {
-                const resolved = resolveNodeValue(prop.value, scope, "", "fetch", fileContent);
-                out[key] = resolved === undefined || resolved === null ? "" : String(resolved);
-            } catch {
-                out[key] = "";
-            }
-        } else if (prop.type === "SpreadElement") {
-            try {
-                const resolved = resolveNodeValue(prop.argument, scope, "", "fetch", fileContent);
-                if (resolved && typeof resolved === "object") {
-                    for (const [k, v] of Object.entries(resolved)) {
-                        if (!(k in out)) out[k] = v === undefined || v === null ? "" : String(v);
-                    }
-                } else {
-                    const chain = memberChainToString(prop.argument);
-                    out[`...${chain ?? "spread"}`] = "<spread>";
-                }
-            } catch {
-                const chain = memberChainToString(prop.argument);
-                out[`...${chain ?? "spread"}`] = "<spread>";
-            }
-        }
-    }
-    return out;
-};
-
-/**
- * Tries to convert a value node into a printable string, transparently
- * unwrapping template literals built from caller-side bindings. Returns null
- * when the value is itself a Vue-style ref / member expression that we can't
- * pin down statically.
- */
-const renderValueNode = (node: any, scope: any, fileContent: string): string | null => {
-    if (!node) return null;
-    try {
-        const resolved = resolveNodeValue(node, scope, "", "fetch", fileContent);
-        if (resolved === null || resolved === undefined) return null;
-        if (typeof resolved === "object") {
-            // Avoid emitting `[object Object]` — leave it to the upstream
-            // placeholder so the consumer knows it's structured.
-            return null;
-        }
-        const s = String(resolved);
-        if (s.startsWith("[unresolved")) return null;
-        return s;
-    } catch {
-        return null;
-    }
-};
-
-/**
- * Substitutes [member:P.X], [param:P], and [urlsearchparams:P.X] markers in a
- * string by walking back to the enclosing function's caller(s). The string is
- * returned with as many markers resolved as we could trace.
- */
-const substituteCallerPlaceholders = (
-    input: string,
-    enclosingFn: EnclosingFn | null,
-    getCallers: (bindingName: string) => CallerInfo[]
-): string => {
-    if (!input || !enclosingFn) return input;
-
-    let output = input;
-
-    output = output.replace(/\[urlsearchparams:([A-Za-z_$][\w$.]*)\]/g, (match, chain: string) => {
-        const parts = chain.split(".");
-        if (parts.length < 1) return match;
-        const paramName = parts[0];
-        const propPath = parts.slice(1);
-        const resolved = resolveParamProperty(paramName, propPath, enclosingFn, getCallers);
-        if (!resolved) return match;
-        const rendered = renderObjectAsQuery(resolved.node, resolved.scope, resolved.fileContent);
-        return rendered ?? match;
-    });
-
-    output = output.replace(/\[member:([A-Za-z_$][\w$.]*)\]/g, (match, chain: string) => {
-        const parts = chain.split(".");
-        if (parts.length < 1) return match;
-        const paramName = parts[0];
-        const propPath = parts.slice(1);
-        const resolved = resolveParamProperty(paramName, propPath, enclosingFn, getCallers);
-        if (!resolved) return match;
-        const rendered = renderValueNode(resolved.node, resolved.scope, resolved.fileContent);
-        return rendered ?? match;
-    });
-
-    return output;
-};
-
-/**
- * Substitutes placeholders in a header bag. `...P.X: <spread>` entries are
- * expanded when the corresponding caller-side value is a literal object;
- * `[member:P.X]` markers inside header values are substituted in-place.
- */
-const substituteCallerHeaders = (
-    headers: Record<string, string>,
-    enclosingFn: EnclosingFn | null,
-    getCallers: (bindingName: string) => CallerInfo[]
-): Record<string, string> => {
-    if (!enclosingFn) return headers;
-    const out: Record<string, string> = {};
-    for (const [k, v] of Object.entries(headers)) {
-        if (k.startsWith("...") && v === "<spread>") {
-            const chain = k.slice(3);
-            const parts = chain.split(".");
-            const paramName = parts[0];
-            const propPath = parts.slice(1);
-            const resolved = resolveParamProperty(paramName, propPath, enclosingFn, getCallers);
-            if (resolved && resolved.node.type === "ObjectExpression") {
-                const obj = renderObjectExpression(resolved.node, resolved.scope, resolved.fileContent);
-                if (obj) {
-                    for (const [hk, hv] of Object.entries(obj)) {
-                        if (!(hk in out)) out[hk] = hv;
-                    }
-                    continue;
-                }
-            }
-            out[k] = v;
-        } else {
-            out[k] = substituteCallerPlaceholders(v, enclosingFn, getCallers);
-        }
-    }
-    return out;
-};
 
 /**
  * Scans all JS files in the given directory for fetch() calls,
@@ -489,77 +178,7 @@ const vue_resolveFetch = async (directory: string, frameworkName = "Vue.JS"): Pr
         // fileAst and fileContent go out of scope here — GC can reclaim them.
     }
 
-    // Caller lookup — searches fetchFilePaths on demand, no persistent cache.
-    //
-    // The previous design used a callerCache (Map<bindingName, CallerInfo[]>) so
-    // repeated lookups for the same name were free. The problem: CallerInfo holds
-    // live Babel AST node references (callNode, scope, args). A cached CallerInfo
-    // keeps the entire file AST alive for the lifetime of vue_resolveFetch, which
-    // can be hundreds of large parsed ASTs simultaneously → OOM.
-    //
-    // Without caching, each getCallers call is self-contained: the CallerInfo
-    // objects are returned, used immediately in resolveParamProperty, and then
-    // become garbage — the AST node refs they hold are released promptly.
-    //
-    // Short (≤2 char) binding names are minifier locals that match thousands of
-    // call sites; they always overflow MAX_CALLERS_PER_NAME and return [] anyway,
-    // so we short-circuit before touching any file.
-    const MAX_CALLERS_PER_NAME = 64;
-    const getCallers = (bindingName: string): CallerInfo[] => {
-        if (!bindingName || bindingName.length <= 2) return [];
-        const needle = `${bindingName}(`;
-        const out: CallerInfo[] = [];
-        let overflowed = false;
-        for (const filePath of fetchFilePaths) {
-            if (overflowed) break;
-            let fileContent: string;
-            try {
-                fileContent = fs.readFileSync(filePath, "utf-8");
-            } catch {
-                continue;
-            }
-            if (!fileContent.includes(needle)) continue;
-            let fileAst: any;
-            try {
-                fileAst = parser.parse(fileContent, {
-                    sourceType: "unambiguous",
-                    plugins: ["jsx", "typescript"],
-                    errorRecovery: true,
-                });
-            } catch {
-                continue;
-            }
-            traverse(fileAst, {
-                CallExpression(callPath) {
-                    if (overflowed) {
-                        callPath.stop();
-                        return;
-                    }
-                    const callee = callPath.node.callee;
-                    if (callee.type !== "Identifier" || callee.name !== bindingName) return;
-                    if (out.length >= MAX_CALLERS_PER_NAME) {
-                        overflowed = true;
-                        callPath.stop();
-                        return;
-                    }
-                    out.push({
-                        file: filePath,
-                        fileContent,
-                        callNode: callPath.node,
-                        scope: callPath.scope,
-                        args: callPath.node.arguments,
-                        enclosingFn: inferEnclosingFn(callPath, filePath),
-                    });
-                },
-            });
-            // fileAst local goes out of scope here. If CallerInfo objects in out[]
-            // hold node refs from this file, those keep the AST alive only until
-            // the caller uses out[] and discards it (no long-lived cache).
-        }
-        // A name with this many callsites is almost certainly a minifier
-        // single-letter local (e, t, n, …) — tracing it is noise, not signal.
-        return overflowed ? [] : out;
-    };
+    const getCallers = makeGetCallers(fetchFilePaths);
 
     let totalFetchCalls = 0;
     // Counter for per-callsite uniqueness so two callsites at the same path+method
