+++ src/map/vue_js/vue_resolveXhr.ts
@@ -5,6 +5,13 @@ import fs from "fs";
 import path from "path";
 import { resolveNodeValue, substituteVariablesInString } from "../next_js/utils.js";
 import * as globals from "../../utility/globals.js";
+import {
+    EnclosingFn,
+    inferEnclosingFn,
+    substituteCallerPlaceholders,
+    substituteCallerHeaders,
+    makeGetCallers,
+} from "./taint_utils.js";
 
 const traverse = _traverse.default;
 
@@ -16,6 +23,7 @@ interface XhrEntry {
     method: string;
     headers: Record<string, string>;
     body: string;
+    enclosingFn: EnclosingFn | null;
 }
 
 /**
@@ -26,6 +34,10 @@ interface XhrEntry {
  * The resolution strategy tracks every binding assigned a `new XMLHttpRequest()`
  * instance and then collects the corresponding .open(), .setRequestHeader(), and
  * .send() calls on that same binding within the enclosing function scope.
+ *
+ * After the initial AST pass, a taint analysis second pass walks back to each
+ * enclosing function's callers to substitute [param:X], [member:P.X], and
+ * [urlsearchparams:P.X] placeholders in the URL, method, headers, and body.
  *
  * Memory design mirrors vue_resolveFetch: ASTs are parsed per-file and freed
  * immediately after each iteration; no persistent AST cache is maintained.
@@ -44,6 +56,12 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
     files = files
         .filter((f) => f.endsWith(".js") && !f.includes("___subsequent_requests"))
         .filter((f) => !fs.lstatSync(path.join(directory, f)).isDirectory());
+
+    // Build the full path list for caller lookup. XHR wrapper callers may live
+    // in any file (not just those containing XMLHttpRequest), so we search all
+    // JS files with a text pre-filter before parsing.
+    const allFilePaths = files.map((f) => path.join(directory, f));
+    const getCallers = makeGetCallers(allFilePaths);
 
     let totalXhrCalls = 0;
     const entries: XhrEntry[] = [];
@@ -78,8 +96,6 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
         }
 
         // Pass 1: find every binding assigned `new XMLHttpRequest()`.
-        // Maps local variable name → set of scopes where it is bound so we can
-        // match the subsequent .open() / .setRequestHeader() / .send() calls.
         const xhrBindingNames = new Set<string>();
 
         traverse(fileAst, {
@@ -87,13 +103,11 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
                 const callee = p.node.callee;
                 if (callee.type !== "Identifier" || callee.name !== "XMLHttpRequest") return;
 
-                // Pattern: var/let/const x = new XMLHttpRequest()
                 const parent = p.parent;
                 if (parent.type === "VariableDeclarator" && parent.id.type === "Identifier") {
                     xhrBindingNames.add(parent.id.name);
                     return;
                 }
-                // Pattern: x = new XMLHttpRequest()  (assignment expression)
                 if (parent.type === "AssignmentExpression" && parent.left.type === "Identifier") {
                     xhrBindingNames.add(parent.left.name);
                     return;
@@ -101,20 +115,10 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
             },
         });
 
-        if (xhrBindingNames.size === 0) {
-            // AST freed here
-            continue;
-        }
+        if (xhrBindingNames.size === 0) continue;
 
         // Pass 2: for each XHR binding, collect .open(), .setRequestHeader(),
-        // and .send() calls. We group them by binding name. Within the same
-        // enclosing function a binding is typically used for a single request;
-        // if reused for multiple requests we emit one entry per .open() call.
-        //
-        // Data collected per binding:
-        //   openCalls:    [{method, url, line}]
-        //   headers:      {name: value}  (accumulated from all setRequestHeader calls)
-        //   bodies:       [bodyString]   (from .send())
+        // and .send() calls. We group them by binding name; one entry per .open() call.
 
         interface XhrCallData {
             method: string;
@@ -122,16 +126,15 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
             line: number;
             headers: Record<string, string>;
             body: string;
+            enclosingFn: EnclosingFn | null;
         }
 
-        // binding name → list of request snapshots (one per .open() call)
         const xhrCallMap = new Map<string, XhrCallData[]>();
-        // binding name → current accumulator (reset on each .open() call)
-        const xhrAccum = new Map<string, { headers: Record<string, string>; body: string }>();
+        const xhrAccum = new Map<string, { headers: Record<string, string>; body: string; enclosingFn: EnclosingFn | null }>();
 
         for (const name of xhrBindingNames) {
             xhrCallMap.set(name, []);
-            xhrAccum.set(name, { headers: {}, body: "" });
+            xhrAccum.set(name, { headers: {}, body: "", enclosingFn: null });
         }
 
         traverse(fileAst, {
@@ -148,7 +151,6 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
                 const args = p.node.arguments;
 
                 if (methodName === "open") {
-                    // open(method, url[, async])
                     if (args.length < 2) return;
 
                     const rawMethod = resolveNodeValue(args[0], p.scope, "", "fetch", fileContent);
@@ -166,7 +168,11 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
 
                     const line = p.node.loc?.start.line ?? 0;
 
-                    // Flush previous accumulator into the call list, start fresh
+                    // Capture enclosing function (for taint analysis second pass) but
+                    // null out the AST node ref to avoid pinning the file AST in memory.
+                    const rawEnclosingFn = inferEnclosingFn(p, filePath);
+                    const enclosingFn = rawEnclosingFn ? { ...rawEnclosingFn, node: null } : null;
+
                     const accum = xhrAccum.get(obj.name)!;
                     xhrCallMap.get(obj.name)!.push({
                         method,
@@ -174,11 +180,10 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
                         line,
                         headers: { ...accum.headers },
                         body: accum.body,
+                        enclosingFn,
                     });
-                    // Reset accumulator for the next .open() on the same binding
-                    xhrAccum.set(obj.name, { headers: {}, body: "" });
+                    xhrAccum.set(obj.name, { headers: {}, body: "", enclosingFn: null });
                 } else if (methodName === "setRequestHeader") {
-                    // setRequestHeader(name, value)
                     if (args.length < 2) return;
                     const rawName = resolveNodeValue(args[0], p.scope, "", "fetch", fileContent);
                     const rawValue = resolveNodeValue(args[1], p.scope, "", "fetch", fileContent);
@@ -192,7 +197,6 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
                             : String(rawValue ?? "");
                     xhrAccum.get(obj.name)!.headers[headerName] = headerValue;
                 } else if (methodName === "send") {
-                    // send([body])
                     if (args.length > 0) {
                         const rawBody = resolveNodeValue(args[0], p.scope, "", "fetch", fileContent);
                         const bodyStr =
@@ -207,15 +211,12 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
             },
         });
 
-        // Merge accumulated headers/body into each open() entry.
-        // The traverse visitor processes statements in source order, so by the
-        // time we visit .send() the accum already has the correct headers from
-        // all prior .setRequestHeader() calls for this request.  However, because
-        // we push on .open() (before .send() is seen), we need a second merge.
+        // Merge accumulated headers/body into each open() entry. Because we push
+        // on .open() (before .send() is seen), we merge the final accumulator state
+        // into the last entry.
         for (const [bindingName, calls] of xhrCallMap.entries()) {
             const accum = xhrAccum.get(bindingName)!;
             if (calls.length > 0) {
-                // Merge final accumulator state into the last .open() entry
                 const last = calls[calls.length - 1];
                 for (const [k, v] of Object.entries(accum.headers)) {
                     if (!(k in last.headers)) last.headers[k] = v;
@@ -233,11 +234,46 @@ const vue_resolveXhr = async (directory: string, frameworkName = "Vue.JS"): Prom
                     method: call.method,
                     headers: call.headers,
                     body: call.body,
+                    enclosingFn: call.enclosingFn,
                 });
             }
         }
 
         // fileAst and fileContent go out of scope here — GC can reclaim them.
+    }
+
+    // Taint analysis second pass: walk back to each XHR call's enclosing function
+    // callers and substitute [param:X], [member:P.X], and [urlsearchparams:P.X]
+    // placeholders. Also resolves [call:X.toUpperCase()] method placeholders.
+    const MARKER_RE = /\[(urlsearchparams|member|param):/;
+    const headersHaveMarkers = (h: Record<string, string>): boolean => {
+        for (const [k, v] of Object.entries(h)) {
+            if (k.startsWith("...") && v === "<spread>") return true;
+            if (MARKER_RE.test(v)) return true;
+        }
+        return false;
+    };
+
+    for (const entry of entries) {
+        if (
+            entry.enclosingFn?.bindingName &&
+            entry.enclosingFn.bindingName.length > 2 &&
+            (MARKER_RE.test(entry.url) ||
+                MARKER_RE.test(entry.method) ||
+                MARKER_RE.test(entry.body) ||
+                headersHaveMarkers(entry.headers))
+        ) {
+            entry.url = substituteCallerPlaceholders(entry.url, entry.enclosingFn, getCallers);
+            entry.body = substituteCallerPlaceholders(entry.body, entry.enclosingFn, getCallers);
+            entry.headers = substituteCallerHeaders(entry.headers, entry.enclosingFn, getCallers);
+
+            // [call:X.Y.toUpperCase()] → resolve the inner member chain, then uppercase
+            const toUpperMatch = entry.method.match(/^\[call:([\w$.]+)\.toUpperCase\(\)\]$/);
+            if (toUpperMatch) {
+                const sub = substituteCallerPlaceholders(`[member:${toUpperMatch[1]}]`, entry.enclosingFn, getCallers);
+                if (!sub.startsWith("[")) entry.method = sub.toUpperCase();
+            }
+        }
     }
 
     // Output and register with OpenAPI collector
